# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcronIQ Forge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Submit a prompt-driven engineering task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
  /tasks/{id}/approve:
    patch:
      operationId: approveTask
      tags: [tasks]
      summary: Approve a task for execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/reject:
    patch:
      operationId: rejectTask
      tags: [tasks]
      summary: Reject a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /pull-requests:
    get:
      operationId: listPullRequests
      tags: [pullRequests]
      summary: List all pull requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Pull requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PullRequest"
  /pull-requests/{id}:
    get:
      operationId: getPullRequest
      tags: [pullRequests]
      summary: Get a pull request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
  /pull-requests/{id}/merge:
    patch:
      operationId: mergePullRequest
      tags: [pullRequests]
      summary: Merge a pull request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
  /pull-requests/{id}/close:
    patch:
      operationId: closePullRequest
      tags: [pullRequests]
      summary: Close a pull request without merging
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
...
components:
...
  schemas:
...
    Task:
      type: object
      required: [id, projectId, type, status, prompt, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        type:
          type: string
        status:
          type: string
        prompt:
          type: string
        plan:
          type: ["string", "null"]
        filesModified:
          type: array
          items:
            type: string
        buildStatus:
          type: ["string", "null"]
        confidenceScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    TaskInput:
      type: object
      required: [projectId, prompt]
      properties:
        projectId:
          type: integer
        prompt:
          type: string
          minLength: 1
        type:
          type: string
...
    PullRequest:
      type: object
      required: [id, projectId, title, status, branch, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        taskId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        branch:
          type: string
        url:
          type: ["string", "null"]
        filesChanged:
          type: integer
        additions:
          type: integer
        deletions:
          type: integer
        createdAt:
          type: string
        mergedAt:
          type: ["string", "null"]
...